BLADERUNNER: Remove use of C99 math
C++ math functions are overloaded so operate using single-precision
when receiving a float input. The C standard library on FreeMiNT
does not fully support C99 math so use of sqrtf, sinf, etc.
instead of the C++ API does not work.

+++ engines/bladerunner/actor_walk.cpp
@@ -196,8 +196,8 @@ bool ActorWalk::tick(int actorId, float stepDistance, bool inWalkLoop) {
 
 	float angle_rad = _facing / 512.0 * M_PI;
 
-	_current.x += stepDistance * sinf(angle_rad);
-	_current.z -= stepDistance * cosf(angle_rad);
+	_current.x += stepDistance * sin(angle_rad);
+	_current.z -= stepDistance * cos(angle_rad);
 	_current.y = _vm->_scene->_set->getAltitudeAtXZ(_current.x, _current.z, &walkboxFound);
 
 	return false;
